Create Exlivery.Orders.Report module.
Define create/1, a public function that recieve a filename and
generate a csv file with all saved orders.
Define build_order_list/0, a private function that returns all saved
orders as a list of strings.
Define order_string/1, a private function that receive an order and
returns it as a string.
Define item_string/1, a private function that receive an item and
returns it as a string.